tracker/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.6.0
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.27.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
inflection==0.5.1
ipython==8.29.0
ipywidgets==8.1.5
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
Markdown==3.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
openapi-codec==1.3.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
sidecar==0.7.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
swagger_ui_bundle==1.1.0
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
